pyproject.toml
[project]
name = "cliendly-solutions-propelix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-unfold>=0.67.0",
    "djangorestframework>=3.16.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
]
